llm-reverse-eng-poc/requirements.txt
sqlalchemy>=2.0
pandas>=2.0
pyyaml>=6.0
openai>=1.30
tqdm>=4.66
